chore: globToRssItems -> pagesGlobToRssItems
