+++ packages/astro-rss/test/rss.test.js
@@ -1,43 +1,20 @@
-import rss, { globToRssItems } from '../dist/index.js';
+import rss from '../dist/index.js';
 import chai from 'chai';
 import chaiPromises from 'chai-as-promised';
 import chaiXml from 'chai-xml';
+import {
+	title,
+	description,
+	site,
+	phpFeedItem,
+	phpFeedItemWithContent,
+	phpFeedItemWithCustomData,
+	web1FeedItem,
+	web1FeedItemWithContent,
+} from './test-utils.js';
 
 chai.use(chaiPromises);
 chai.use(chaiXml);
-
-const title = 'My RSS feed';
-const description = 'This sure is a nice RSS feed';
-const site = 'https://example.com';
-
-const phpFeedItem = {
-	link: '/php',
-	title: 'Remember PHP?',
-	pubDate: '1994-05-03',
-	description:
-		'PHP is a general-purpose scripting language geared toward web development. It was originally created by Danish-Canadian programmer Rasmus Lerdorf in 1994.',
-};
-const phpFeedItemWithContent = {
-	...phpFeedItem,
-	content: `<h1>${phpFeedItem.title}</h1><p>${phpFeedItem.description}</p>`,
-};
-const phpFeedItemWithCustomData = {
-	...phpFeedItem,
-	customData: '<dc:creator><![CDATA[Buster Bluth]]></dc:creator>',
-};
-
-const web1FeedItem = {
-	// Should support empty string as a URL (possible for homepage route)
-	link: '',
-	title: 'Web 1.0',
-	pubDate: '1997-05-03',
-	description:
-		'Web 1.0 is the term used for the earliest version of the Internet as it emerged from its origins with Defense Advanced Research Projects Agency (DARPA) and became, for the first time, a global network representing the future of digital communications.',
-};
-const web1FeedItemWithContent = {
-	...web1FeedItem,
-	content: `<h1>${web1FeedItem.title}</h1><p>${web1FeedItem.description}</p>`,
-};
 
 // note: I spent 30 minutes looking for a nice node-based snapshot tool
 // ...and I gave up. Enjoy big strings!
@@ -137,98 +114,5 @@ describe('rss', () => {
 		});
 
 		chai.expect(body).xml.to.equal(validXmlResult);
-	});
-});
-
-describe('globToRssItems', () => {
-	it('should generate on valid result', async () => {
-		const globResult = {
-			'./posts/php.md': () =>
-				new Promise((resolve) =>
-					resolve({
-						url: phpFeedItem.link,
-						frontmatter: {
-							title: phpFeedItem.title,
-							pubDate: phpFeedItem.pubDate,
-							description: phpFeedItem.description,
-						},
-					})
-				),
-			'./posts/nested/web1.md': () =>
-				new Promise((resolve) =>
-					resolve({
-						url: web1FeedItem.link,
-						frontmatter: {
-							title: web1FeedItem.title,
-							pubDate: web1FeedItem.pubDate,
-							description: web1FeedItem.description,
-						},
-					})
-				),
-		};
-
-		const items = await globToRssItems(globResult);
-
-		chai.expect(items.sort((a, b) => a.pubDate - b.pubDate)).to.deep.equal([
-			{
-				title: phpFeedItem.title,
-				link: phpFeedItem.link,
-				pubDate: new Date(phpFeedItem.pubDate),
-				description: phpFeedItem.description,
-			},
-			{
-				title: web1FeedItem.title,
-				link: web1FeedItem.link,
-				pubDate: new Date(web1FeedItem.pubDate),
-				description: web1FeedItem.description,
-			},
-		]);
-	});
-
-	it('should fail on missing "url"', () => {
-		const globResult = {
-			'./posts/php.md': () =>
-				new Promise((resolve) =>
-					resolve({
-						url: undefined,
-						frontmatter: {
-							pubDate: phpFeedItem.pubDate,
-							description: phpFeedItem.description,
-						},
-					})
-				),
-		};
-		return chai.expect(
-			rss({
-				title,
-				description,
-				items: globResult,
-				site,
-			})
-		).to.be.rejected;
-	});
-
-	it('should fail on missing "title" key', () => {
-		const globResult = {
-			'./posts/php.md': () =>
-				new Promise((resolve) =>
-					resolve({
-						url: phpFeedItem.link,
-						frontmatter: {
-							title: undefined,
-							pubDate: phpFeedItem.pubDate,
-							description: phpFeedItem.description,
-						},
-					})
-				),
-		};
-		return chai.expect(
-			rss({
-				title,
-				description,
-				items: globResult,
-				site,
-			})
-		).to.be.rejected;
 	});
 });
